receipt-app/requirements.txt
streamlit>=1.30
Pillow>=10.0
pandas>=2.0
python-dateutil>=2.8.2
openpyxl>=3.1
easyocr>=1.7
numpy
opencv-python-headless>=4.8
anthropic>=0.30
supabase>=2.0
